feat: restructure components by moving Header and adding TaskList
