+++ week9/vite-project/src/components/TaskList.jsx
@@ -1,0 +1,31 @@
+import React from "react";
+
+const tasks = [
+  {
+    id: 1,
+    title: "Review week 9 material",
+    date: "June 4th at 1 pm",
+  },
+  {
+    id: 2,
+    title: "Do quiz 9",
+    date: "June 4th at 6 pm",
+  },
+  {
+    id: 3,
+    title: "Work on assignment 2",
+    date: "June 5th at 8 am",
+  },
+];
+
+function TaskList() {
+  return (
+    <ul>
+      {tasks.map((task) => (
+        <li key={task.id}>{task.title}</li>
+      ))}
+    </ul>
+  );
+}
+
+export default TaskList;
